申請フォームエディター周りのルーティングファイルや名前空間を整理
